Sponsor page: page move to a client component
